mhbdips_server/requirements.txt
asgiref==3.8.1
Django==4.2.11
django-rest-framework==0.1.0
djangorestframework==3.15.1
future==1.0.0
http-request==1.2
pillow==10.2.0
psycopg2==2.9.9
sqlparse==0.4.4
typing_extensions==4.10.0
